docs(config): move config into file and add link
